modularise scripts
