+++ src/scripts/vscode.ts
@@ -1,0 +1,5 @@
+import { provideVSCodeDesignSystem, vsCodeButton, vsCodeCheckbox, vsCodeDivider, vsCodeTextField } from '@vscode/webview-ui-toolkit';
+
+provideVSCodeDesignSystem().register(vsCodeButton(), vsCodeCheckbox(), vsCodeTextField(), vsCodeDivider());
+
+export const vscode = acquireVsCodeApi();
